+++ Budget/DayBudget/DayBudgetManager.swift
@@ -35,51 +35,24 @@ final class DayBudgetManager {
         return dayBudget
     }
 
-    public func budget() -> Int {
+    public var budget: Int {
         return dayBudget.budget
     }
 
-    public func balance() -> Int  {
+    public var balance: Int  {
         return dayBudget.balance
     }
 
-    public func date() -> Date  {
+    public var date: Date {
         return dayBudget.date
     }
 
-
-    public func dateString() -> String  {
+    public var dateString: String  {
         return DayBudget.dateFormat.string(from: dayBudget.date)
     }
 
-    public func addExpense(_ expense: Int) {
+    public func add(expense: Int) {
         dayBudget.addExpense(expense)
         persistentManager.storeDayBudget(dayBudget)
-    }
-}
-
-protocol PersistDayBudgetStrategy {
-    func getDayBudget() -> DayBudget?
-    func storeDayBudget(_ dayBudget: DayBudget)
-}
-
-class UserDefaultsManager: PersistDayBudgetStrategy {
-    static let balanceKey = "BalanceKey"
-    static let dateKey = "DateKey"
-
-    let defaults = UserDefaults.standard
-
-    func getDayBudget() -> DayBudget? {
-        guard let balance = defaults.value(forKey: UserDefaultsManager.balanceKey) as? Int,
-                let date = defaults.value(forKey: UserDefaultsManager.dateKey) as? Date else {
-                return nil
-        }
-
-        return DayBudget(budget: balance, date: date)
-    }
-
-    func storeDayBudget(_ dayBudget: DayBudget) {
-        defaults.set(dayBudget.balance, forKey: UserDefaultsManager.balanceKey)
-        defaults.set(dayBudget.date, forKey: UserDefaultsManager.dateKey)
     }
 }
